use int32 instead

+++ onnx/backend/test/case/node/negativeloglikelihoodloss.py
@@ -28,7 +28,7 @@ def compute_negative_log_likelihood_loss(input, target, weight=None, reduction='
         # when the target value is > C or < 0, it doesn't matter which value we are
         # taking in gather_weight, since it will be set to 0 in the following if-block
         # use np.int_ to make it compatible with x86 machines
-        gather_weight = np.take(weight, np.array(target, dtype=np.int_), mode='clip')
+        gather_weight = np.take(weight, np.array(target, dtype=np.int32), mode='clip')
         # set `ignore_index`'s loss weight to 0.
         # The loss tensor will be multiplied by this weight tensor,
         # so `ingore_index`'s loss value will be eliminated.
